mutatomic

Tools for Safe Pass-By-Reference Modification Semantics on Atomic Objects

What is 'mutatomic'?

'mutatomic' is an 'R' package that provides 2 things:

For regular 'R' users:
A new class of atomic vectors, matrices, and arrays, called mutatomic, that provides (safer) support for pass-by-reference semantics.
It also provides an optional helper function, called currentBindings.

For package back-end:
A set of tools for safer pass-by-reference modification semantics for atomic objects, as 'R' does not natively provide pass-by-reference mechanics (except via it's internal C API or through a package like 'Rcpp').
The developer tools provided by 'mutatomic' were primarily designed for the 'squarebrackets' and 'broadcast' packages, but authors of other packages are welcome to use 'mutatomic' also.

'mutatomic' does not come with pass-by-reference functions itself, except for 2 simple example functions (ma_set and setapply).

For Regular 'R' Users

Regular 'R' users can construct objects of class 'mutatomic', or convert objects to class 'mutatomic';
see mutatomic_class.

Depending on the situation end users may also use the currentBindings function.

The 'mutatomic' package also comes with a few help pages, that end users can refer to, to gain a better understanding of the pass-by-reference semantics supported by 'mutatomic'.
These help pages are the following:

mutatomic_class:
Explains the 'mutatomic' class.

mutatomic_PassByReference:
Explains Pass-by-Reference semantics, and its important consequences.

mutatomic_coercion:
Explains the difference in coercion rules between modification through Pass-by-Reference semantics and modification through copy (i.e. pass-by-value).

For Package Back-End

'mutatomic' provides tools for package authors to program with mutable atomic objects.
Arguably the most important function in 'mutatomic' for developers is the stopifnot_ma_safe2mutate function.
This function checks if an atomic object is safe to mutate, and gives an error otherwise.

What follows is technical information on 'mutatomic', and why a 'mutatomic' class is needed; this is of no interest for regular 'R' users.

Technical - Why is 'mutatomic' needed?

Consider the following code:

x <- base::letters
collapse::setv(x, "a", "xxx")

the above code modifies base::letters by reference, and nothing is stopping the user from changing base 'R' while 'R' is still running!
Now, obviously collapse::setv() was meant for professional programmers, and not to be called by amateurs.
But what if one wishes to design an 'R' package that provides pass-by-reference mechanics for atomic objects in a safer way?

This is where the 'mutatomic' package comes in.
'mutatomic' provides a new class which can be considered the atomic version of the 'data.table' class, and through this class package authors can protect the user from changing things like base:letters.

Technical - How does 'mutatomic' solve the issue?

'mutatomic' first and fore-most provides a new class, called 'mutatomic'.
If a function demands an atomic object is of class 'mutatomic' to allow pass-by-reference semantics, the above issue is (mostly) prevented.
The issue is prevented through the following means:

Creating an object of class 'mutatomic', at least when using the functions provided by this package, will copy the original object.
So when calling x <- mutatomic(base::letters), x no longer refers to base::letters.

When a function demands an object is of class 'mutatomic', it is (almost) guaranteed it does not refer to an base 'R' object, so something like base::letters will never be modified by reference.

'mutatomic' stores a list of most base 'R' (atomic) object addresses when it is loaded.
The is.mutatomic and stopifnot_ma_safe2mutate functions check this list, creating extra certainty that base 'R' is never modified by reference.

The 'mutatomic' class is not just defined by a class name attribute.
The class comes with additional attributes to ensure it truly has been created by the functions of this package.
Although these attributes can be mimicked with enough effort, it is very unlikely for an object to accidentally have these attributes.
The is.mutatomic and stopifnot_ma_safe2mutate functions check for these attributes, creating extra security.

'mutatomic' respects the lock of a binding, if there is any.

References

The badges shown in the documentation of this R-package were made using the services of: https://shields.io/

Author

Author, Maintainer: Tony Wilkes tony_a_wilkes@outlook.com (ORCID)